backend/requirements.txt
google-genai>=1.30.0
psycopg2-binary>=2.9.9
pydantic>=2.11.0
python-dotenv>=1.0.1
requests>=2.32.3
